% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetframe.R
\name{frameOptions}
\alias{frameOptions}
\title{Options for widget's iframe.}
\usage{
frameOptions(xdomain = "*", title = NULL, name = NULL, id = NULL,
  allowfullscreen = FALSE, sandbox = NULL, lazyload = FALSE)
}
\arguments{
\item{xdomain}{xdomain to validate messages received.}

\item{title}{If passed it will be assigned to the iframe title attribute.}

\item{name}{If passed it will be assigned to the iframe name attribute.}

\item{id}{If passed it will be assigned to the iframe id attribute.}

\item{allowfullscreen}{If TRUE it will set the iframe allowfullscreen attribute to true.}

\item{sandbox}{If passed it will be assigned to the iframe sandbox attribute.}

\item{lazyload}{If TRUE the child widget is lazy loaded using \href{http://dinbror.dk/blog/blazy/?ref=github#iframe}{bLazy.js}.}
}
\description{
Taken from \href{http://blog.apps.npr.org/pym.js/api/pym.js/1.1.2/module-pym.Parent.html}{Pym.js Documentation}.
In addition also check out the \href{https://developer.mozilla.org/en-US/docs/Web/HTML/Element/iframe}{iframe documentation}.
}
