% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetframe.R
\name{saveWidgetframe}
\alias{saveWidgetframe}
\title{Save a widgetframe and its child widget to HTML files}
\usage{
saveWidgetframe(widget, file, selfcontained = FALSE, libdir = NULL,
  background = "white", knitrOptions = list())
}
\arguments{
\item{widget}{widgetframe to save}

\item{file}{File to save  the parent HTML into. The child HTML will be saved to `basename(file)_widget/index.html`.}

\item{selfcontained}{Whether to save the parent and child HTML as a single self-contained files.
(with external resources base64 encoded) or files with external resources
placed in an adjacent directory.}

\item{libdir}{Directory to copy HTML dependencies into (defaults to
filename_files).}

\item{background}{Text string giving the html background color of the widget.
Defaults to white.}

\item{knitrOptions}{A list of \pkg{knitr} chunk options.}
}
\description{
Save a rendered widgetframe and its child widget to HTML files.
}

