% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetframe-package.r
\docType{package}
\name{widgetframe}
\alias{widgetframe}
\alias{widgetframe-package}
\title{widgetframe: A package for wrapping htmlwidgets in responsive iframes.}
\description{
This package provides two functions \code{\link{frameableWidget}}, and \code{\link{frameWidget}}.
The \code{\link{frameableWidget}} is used to add extra code to a htmlwidget which allows is to be rendered inside a responsive iframe.
The \code{\link{frameWidget}} is a htmlwidget which displays content of another htmlwidget inside a responsive iframe.
}

