% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetframe-package.r, R/widgetframe.R
\docType{package}
\name{widgetframe}
\alias{widgetframe}
\alias{widgetframe-package}
\title{widgetframe.}
\usage{
widgetframe(url, width = "100\%", height = NULL, elementId = NULL)
}
\description{
widgetframe.

<Add Description>
}

