% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetframe-package.r
\docType{package}
\name{widgetframe}
\alias{widgetframe}
\alias{widgetframe-package}
\title{widgetframe: A package for wrapping htmlwidgets in responsive iframes.}
\description{
This pacakge provides two functions `frameableWidget`, and `frameWidget`.
The `frameableWidget` is used to add extra code to a htmlwidget which allows is to be rendered inside a responsive iframe.
The `frameWidget` is a htmlwidget which displays content of another htmlwidget inside a responsive iframe. These functions allow for easier embedding of htmlwidgets in CMS systems such as wordpress, blogger etc. They also allow for seperation of widget content from main HTML content.
}

